gRPC/requirements.txt
grpcio>=1.59.2
fastapi>=0.104.1
uvicorn>=0.24
